+++ spacy/tests/tokenizer/test_tokenizer.py
@@ -1,9 +1,43 @@
-import pytest
 import re
-from spacy.vocab import Vocab
-from spacy.tokenizer import Tokenizer
-from spacy.util import ensure_path, compile_prefix_regex, compile_suffix_regex
+
+import pytest
 from spacy.lang.en import English
+from spacy.tokenizer import Tokenizer
+from spacy.tokens import Doc
+from spacy.util import compile_prefix_regex, compile_suffix_regex, ensure_path
+from spacy.vocab import Vocab
+
+
+@pytest.mark.issue(743)
+def test_issue743():
+    doc = Doc(Vocab(), ["hello", "world"])
+    token = doc[0]
+    s = set([token])
+    items = list(s)
+    assert items[0] is token
+
+
+@pytest.mark.issue(801)
+@pytest.mark.skip(
+    reason="Can not be fixed unless with variable-width lookbehinds, cf. PR #3218"
+)
+@pytest.mark.parametrize(
+    "text,tokens",
+    [
+        ('"deserve,"--and', ['"', "deserve", ',"--', "and"]),
+        ("exception;--exclusive", ["exception", ";--", "exclusive"]),
+        ("day.--Is", ["day", ".--", "Is"]),
+        ("refinement:--just", ["refinement", ":--", "just"]),
+        ("memories?--To", ["memories", "?--", "To"]),
+        ("Useful.=--Therefore", ["Useful", ".=--", "Therefore"]),
+        ("=Hope.=--Pandora", ["=", "Hope", ".=--", "Pandora"]),
+    ],
+)
+def test_issue801(en_tokenizer, text, tokens):
+    """Test that special characters + hyphens are split correctly."""
+    doc = en_tokenizer(text)
+    assert len(doc) == len(tokens)
+    assert [t.text for t in doc] == tokens
 
 
 def test_tokenizer_handles_no_word(tokenizer):
